extract common layout components for apps/appset
